[CBRD-25352] When adding a user schema to a stored procedure, the way a built-in PL/SQL is called is changed, and the test tools need to be temporarily modified. (#681)
http://jira.cubrid.org/browse/CBRD-25352

When calling Built-in PL/CSQL in phase-1, “dbms_output.” must be added for normal execution.

When that PR is merged, it will affect the develop branch where the existing phase-0 was merged.
Therefore, we modified it to perform the built-in PL/CSQL once more by temporarily removing the 'DBMS_OUTPUT." inside the catch statement until Phase-1 is merged.

We plan to remove it again when the phase-1 feature branch is merged in the future.